Adaptions: create dict_esmtimestamp-label which maps the ESM_timestamps to the label_human motion - general labels. This is needed in order to be able to use the feature selection function from iteration 01.
ADDITIONALLY, discovered and solved an error in the labeling of the sensor data: the initial label information was stored in dictionary (one for every participant) with the keys being the activities. If one participant collected two times the same activiy, the first activity information was overriten. This problem is now solved.

Note: it took SOO much longer than expected...